reset conflict cli model
