+++ php_modules/devtool/starter/controllers/database.php
@@ -3,58 +3,9 @@ namespace App\devtool\starter\controllers;
 
 use SPT\Response;
 use SPT\Web\ControllerMVVM;
-
-class cli extends ControllerMVVM
+ 
+class database extends ControllerMVVM
 {
-    public function list()
-    {
-        $list = $this->StarterModel->getSolutions();
-        foreach($list as $item)
-        {
-            echo $item->name." : ". $item->description. "\n";
-        }
-
-        return true;
-    }
-
-    public function install()
-    {
-        $args = $this->request->cli->getArgs();
-        $solution = isset($args[1]) ? $args[1] : '';
-
-        if (file_exists($solution))
-        {
-            $try = $this->StarterModel->installFileZipCli($solution);
-        } else {
-            $try = $this->StarterModel->install($solution);
-        }
-        
-        if ($try['success'])
-        {
-            echo "Install Done!\n";
-        }
-
-        return true;
-    }
-
-    public function uninstall()
-    {
-        $args = $this->request->cli->getArgs();
-        $solution = isset($args[1]) ? $args[1] : '';
-
-        $try = $this->StarterModel->uninstall($solution);
-        if (!$try)
-        {
-            echo $this->StarterModel->getError() ."\n";
-        }
-        else
-        {
-            echo "Uninstall Done!\n";
-        }
-
-        return true;
-    }
-
     public function checkavailability()
     {
         $entities = $this->DbToolModel->getEntities();
